useless comment of a mega commit

+++ cc.sh
@@ -8,12 +8,13 @@ app=$(basename $PWD)
 test -e dist || (mkdir dist && mkdir dist/build && mkdir dist/build/$app)
 
 while true; do
+  printf '\033\143' #clear
   nix-shell --run "cabal configure --ghcjs && time cabal build $app $@; true"
 
   # used by ghcjsi repl
-  sed -i 's/h$main(h$mainZCZCMainzimain);/module.exports = { h$main: h$main, h$killThread: h$killThread, h$d: h$d, h$baseZCControlziExceptionziBasezinonTermination: h$baseZCControlziExceptionziBasezinonTermination };h$main(h$mainZCZCMainzimain);/g' dist/build/$app/$app.jsexe/all.js
+  test -f dist/build/$app/$app.jsexe/all.js && sed -i 's/h$main(h$mainZCZCMainzimain);/module.exports = { h$main: h$main, h$killThread: h$killThread, h$d: h$d, h$baseZCControlziExceptionziBasezinonTermination: h$baseZCControlziExceptionziBasezinonTermination };h$main(h$mainZCZCMainzimain);/g' dist/build/$app/$app.jsexe/all.js
 
-  cp -f dist/build/$app/$app.jsexe/all.js $app/
+  test -f dist/build/$app/$app.jsexe/all.js && cp -f dist/build/$app/$app.jsexe/all.js $app/
   nix-shell -p fswatch --run "fswatch -1 -r -i '.*[.]hs$' --event Created --event Updated --event Removed --event Renamed --event MovedFrom --event MovedTo src; true"
 done
